refactor: Move client disposal and request data processing logic to cleanup module for better organization
