+++ api/server/controllers/AskController.js
@@ -1,108 +1,18 @@
 const { getResponseSender, Constants } = require('librechat-data-provider');
 const {
+  handleAbortError,
   createAbortController,
   cleanupAbortController,
-  handleAbortError,
 } = require('~/server/middleware');
+const {
+  disposeClient,
+  processReqData,
+  clientRegistry,
+  requestDataMap,
+} = require('~/server/cleanup');
 const { sendMessage, createOnProgress } = require('~/server/utils');
 const { saveMessage } = require('~/models');
 const { logger } = require('~/config');
-
-const requestDataMap = new WeakMap();
-const FinalizationRegistry = global.FinalizationRegistry || null;
-
-const clientRegistry = FinalizationRegistry
-  ? new FinalizationRegistry((heldValue) => {
-    try {
-      // This will run when the client is garbage collected
-      if (heldValue && heldValue.abortKey) {
-        cleanupAbortController(heldValue.abortKey);
-      }
-    } catch (e) {
-      // Ignore errors
-    }
-  })
-  : null;
-
-function disposeClient(client) {
-  if (!client) {
-    return;
-  }
-  try {
-    if (client.apiKey) {
-      client.apiKey = null;
-    }
-    if (client.azure) {
-      client.azure = null;
-    }
-    if (client.sendMessage) {
-      client.sendMessage = null;
-    }
-    if (client.currentMessages) {
-      client.currentMessages = null;
-    }
-    if (client.getBuildMessages) {
-      client.getBuildMessages = null;
-    }
-    if (client.getResponseSender) {
-      client.getResponseSender = null;
-    }
-    if (client.currentMessages) {
-      client.currentMessages = null;
-    }
-    if (client.streamHandler) {
-      client.streamHandler = null;
-    }
-    if (client.abortController) {
-      client.abortController = null;
-    }
-    if (client.options) {
-      client.options = null;
-    }
-    if (client.savedMessageIds) {
-      client.savedMessageIds.clear();
-      client.savedMessageIds = null;
-    }
-    if (typeof client.dispose === 'function') {
-      client.dispose();
-    }
-  } catch (e) {
-    // Ignore
-  }
-}
-
-function processReqData(data = {}, context) {
-  let {
-    userMessage,
-    userMessagePromise,
-    responseMessageId,
-    promptTokens,
-    conversationId,
-    userMessageId,
-  } = context;
-  for (const key in data) {
-    if (key === 'userMessage') {
-      userMessage = data[key];
-      userMessageId = data[key].messageId;
-    } else if (key === 'userMessagePromise') {
-      userMessagePromise = data[key];
-    } else if (key === 'responseMessageId') {
-      responseMessageId = data[key];
-    } else if (key === 'promptTokens') {
-      promptTokens = data[key];
-    } else if (!conversationId && key === 'conversationId') {
-      conversationId = data[key];
-    }
-  }
-  return {
-    userMessage,
-    userMessagePromise,
-    responseMessageId,
-    promptTokens,
-    conversationId,
-    userMessageId,
-  };
-}
 
 const AskController = async (req, res, next, initializeClient, addTitle) => {
   let {
@@ -153,6 +63,7 @@ const AskController = async (req, res, next, initializeClient, addTitle) => {
 
   const updateReqData = (data = {}) => {
     reqDataContext = processReqData(data, reqDataContext);
+    abortKey = reqDataContext.abortKey;
     userMessage = reqDataContext.userMessage;
     userMessagePromise = reqDataContext.userMessagePromise;
     responseMessageId = reqDataContext.responseMessageId;
@@ -178,6 +89,7 @@ const AskController = async (req, res, next, initializeClient, addTitle) => {
     }
 
     if (abortKey) {
+      logger.debug('[AskController] Cleaning up abort controller');
       cleanupAbortController(abortKey);
       abortKey = null;
     }
@@ -232,12 +144,12 @@ const AskController = async (req, res, next, initializeClient, addTitle) => {
       };
     };
 
-    const {
-      abortController,
-      onStart,
-      abortKey: _aK,
-    } = createAbortController(req, res, getAbortData, updateReqData);
-    abortKey = _aK;
+    const { onStart, abortController } = createAbortController(
+      req,
+      res,
+      getAbortData,
+      updateReqData,
+    );
 
     const closeHandler = () => {
       logger.debug('[AskController] Request closed');
@@ -320,19 +232,11 @@ const AskController = async (req, res, next, initializeClient, addTitle) => {
       });
     }
 
-    if (
-      typeof addTitle === 'function' &&
-      parentMessageId === Constants.NO_PARENT &&
-      newConvo &&
-      !abortController.signal.aborted
-    ) {
-      const titleResponse = { ...response };
-
-      // Pass the full client object as requested
+    if (typeof addTitle === 'function' && parentMessageId === Constants.NO_PARENT && newConvo) {
       addTitle(req, {
         text,
-        response: titleResponse,
-        client: client, // Pass the actual client object
+        response: { ...response },
+        client,
       })
         .then(() => {
           logger.debug('[AskController] Title generation started');
