+++ api/server/routes/ask/custom.js
@@ -3,7 +3,6 @@ const AskController = require('~/server/controllers/AskController');
 const { initializeClient } = require('~/server/services/Endpoints/custom');
 const { addTitle } = require('~/server/services/Endpoints/openAI');
 const {
-  handleAbort,
   setHeaders,
   validateModel,
   validateEndpoint,
